use separate views file
